Reduce resource usage in browser environments by using different options
